fixed: change function to the outside of utils
